多线程版本，注意多线程中无法注入bean问题的解决，以及aop，等

+++ src/main/java/com/arviiin/dataquality/mapper/RolesMapper.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Mapper
 public interface RolesMapper {
 
-    int addRoles(@Param("roles") String[] roles, @Param("uid") Long uid);
+    int addRoles(@Param("roles") String[] roles, @Param("uid") Integer uid);
 
     List<Role> getRolesByUid(Long uid);
 }
